+++ Cargo.toml
@@ -9,12 +9,12 @@ repository = "https://github.com/denoland/rustls-tokio-stream"
 
 [dependencies]
 tokio = { version = "1", features = [ "io-util", "net", "rt" ] }
-rustls = "0.20.5"
-futures = "0.3.21"
-parking_lot = "0.12.0"
+rustls = "0.20"
+futures = "0.3"
+parking_lot = "0.12"
 
 [dev_dependencies]
 tokio = { version = "1", features = [ "full" ] }
-rustls-pemfile = "1.0.0"
-rustls = { version = "0.20.5", features = [ "dangerous_configuration" ] }
+rustls-pemfile = "1.0"
+rustls = { version = "0.20", features = [ "dangerous_configuration" ] }
 ntest = "0.9"
